Exercise/src/Product.java
public class Product {
    private int reference;
    private String name;
    private double price;

    public Product(int reference, String name, double price) {
        this.reference = reference;
        this.name = name;
        this.price = price;
    }

    public int getReference() {
        return reference;
    }

    public void setReference(int reference) {
        this.reference = reference;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public double getPrice() {
        return price;
    }

    public void setPrice(double price) {
        this.price = price;
    }
}
